requirements.txt
flask
flasgger
flask_restful
pyairtable
gunicorn
firecrawl-py
pydantic